requirements.txt
fastapi
uvicorn
playwright
httpx
websockets
pydantic
aiofiles
python-dotenv
slowapi
